unitpriceを整数型に、mileageをfloatに

+++ src/models/fuel_efficiency.rs
@@ -7,8 +7,8 @@ pub struct FuelEfficiency {
     pub car_id: i32,
     pub fe_date: String,
     pub fe_amount: f32,
-    pub fe_unitprice: f32,
-    pub fe_mileage: i32,
+    pub fe_unitprice: i32,
+    pub fe_mileage: f32,
     pub created_at: Option<OffsetDateTime>,
     pub updated_at: Option<OffsetDateTime>,
 }
